+++ CHANGELOG.md
@@ -2,8 +2,15 @@
 
 ## Unreleased
 
+## v0.1.1a1
+
 - (docs) Switch to manual API listing.
 - (docs) Pin `sphinx-autodoc2` to our own fork before its [#17](https://github.com/sphinx-extensions2/sphinx-autodoc2/issues/17) is fixed.
+- Add timeout to `run_executable` since some commands may require human's input and get stuck.
+- Add `grading.repository.RepositoryBaseTestCase`.
+- Add `Repository.get_all_tag_refs`.
+- Add `Repository.get_tag_refs_at`.
+- Fix `Repository` from an archive file is using the archive's filename instead of the "repo".
 
 ## v0.1.0